requirements.txt
sympy
numpy
pyyaml
pretty_midi
mido
pyfluidsynth
pydub
music21
soundfile
scipy
pytest
audioop-lts
nicegui
pyinstaller